Made useForm from useEdit but with validation too
