+++ src/utils/mixins/editMixin.js
@@ -1,43 +1,7 @@
 import cloneDeep from 'clone-deep'
 import deepEqual from 'deep-equal'
-import { computed, ref, unref, watch } from 'vue'
 
 import { objectDiff } from '@/utils/utils'
-
-export function useEdit (initial) {
-  const edit = ref(cloneDeep(unref(initial)))
-
-  const isNew = computed(() => {
-    return initial.value && !initial.value.id && initial.value.id !== 0
-  })
-
-  const hasChanged = computed(() => {
-    return !isNew.value && !deepEqual(initial.value, edit.value)
-  })
-
-  watch(initial, (current, previous) => {
-    const changes = current && previous ? objectDiff(previous, current) : {}
-    edit.value = Object.assign(edit.value, cloneDeep(changes))
-  })
-
-  const saveData = computed(() => {
-    return isNew.value ? edit.value : { id: initial.value.id, ...objectDiff(initial.value, edit.value) }
-  })
-
-  function reset () {
-    edit.value = cloneDeep(initial.value)
-    // TODO: is this needed?
-    // this.$emit('reset', this.value.id)
-  }
-
-  return {
-    edit,
-    saveData,
-    isNew,
-    hasChanged,
-    reset,
-  }
-}
 
 export default {
   props: {
